V1

+++ src/views/asideContent/Customer/CustomerManagementContent.vue
@@ -1,5 +1,4 @@
 <template>
-  <!-- TODO:2023-5-9 这一天请你个给前端github远程分支添加dev -->
   <div id="Order">
     <div class="OrderList">
       <div class="top-operation">
@@ -46,15 +45,9 @@
         </div>
       </div>
       <div class="table">
-        <!-- TODO:高度需改成动态，以便响应式 -->
         <el-table
           ref="multipleTableRef"
-          :data="
-            filterData.slice(
-              (currentPage - 1) * pageSize,
-              currentPage * pageSize
-            )
-          "
+          :data="filterData"
           stripe
           style="width: 100%"
           border
@@ -97,7 +90,7 @@
             v-model:page-size="pageSize"
             :page-sizes="[5, 10, 20, 50]"
             layout="sizes, prev, pager, next, jumper"
-            :total="tableData.length"
+            :total="pageTotal"
             @size-change="handleSizeChange"
             @current-change="handleCurrentChange"
           />
@@ -109,32 +102,36 @@
   <el-dialog v-model="dialogInfo.isShow" :title="dialogInfo.title">
     <CustomerManagementDialog :updateData="getTableData" />
   </el-dialog>
-
-  <deleteDialog :updateData="getTableData" />
 </template>
 
 <script lang="ts" setup>
-import { ref, onMounted, watchEffect, reactive, defineEmits } from "vue";
+import { ref, onMounted, watchEffect, reactive } from "vue";
 import { debounce } from "lodash"; // 引入防抖方法
+import type { Action } from "element-plus";
 import { Refresh, User } from "@element-plus/icons-vue";
-import { ElConfigProvider, ElTable, ElNotification } from "element-plus";
+import {
+  ElConfigProvider,
+  ElTable,
+  ElNotification,
+  ElMessage,
+  ElMessageBox,
+} from "element-plus";
 import { useDialogStore } from "../../../store/store";
 import { storeToRefs } from "pinia";
 import zhCn from "element-plus/lib/locale/lang/zh-cn"; // 引入中文包
 import { Search, Plus, EditPen, CloseBold } from "@element-plus/icons-vue";
 import ApiClient from "../../../request/request";
 import CustomerManagementDialog from "../../../components/formContent/CustomerManagement.vue";
-import deleteDialog from "../../../components/confirmDelete.vue";
 import type {
   apiResponseCustomerRepresentative,
   apiResponseData,
   apiResponseUser,
 } from "../../../model/interface";
-import type { CustomerRepresentative } from "../../../model/users";
+import type { CustomerRepresentativeInfo } from "../../../model/users";
 const emit = defineEmits(["updateData"]);
 const apiClient = ApiClient.getInstance();
 const dialogStore = useDialogStore();
-const { dialogInfo, confirmDelete, delUrl }: any = storeToRefs(dialogStore);
+const { dialogInfo }: any = storeToRefs(dialogStore);
 // 搜索方式接口
 interface options {
   value: string;
@@ -162,18 +159,28 @@ const loading = ref<boolean>(true);
 // 存储搜索框内容
 const searchValue = ref<string>("");
 // 存储表单数据
-const tableData = ref<CustomerRepresentative[]>([]);
+const tableData = ref<CustomerRepresentativeInfo[]>([]);
 // 存储搜索关键字过滤后的数据
-const filterData = ref<CustomerRepresentative[]>([]);
+const filterData = ref<CustomerRepresentativeInfo[]>([]);
+// 分页框 页数
+const currentPage = ref<number>(1);
+// 显示表单数据行数
+const pageSize = ref<number>(10);
+// 表单总数
+const pageTotal = ref<number>(0);
 // 获取表单数据
 const getTableData = async (): Promise<void> => {
   try {
+    searchValue.value = "";
     loading.value = true;
+    // /CustomerRepresentative?pageSize=${数据大小}&page=${页数}
     const response = await apiClient.get<apiResponseCustomerRepresentative>(
-      "/CustomerRepresentative"
+      `/CustomerRepresentative?pageSize=${pageSize.value}&page=${currentPage.value}`
     );
-    const responseData: any = response!.data;
-    tableData.value = responseData.map((user: CustomerRepresentative) => ({
+    const responseData: any = response!.data.data;
+    // 页面数据长度
+    pageTotal.value = response!.data.total;
+    tableData.value = responseData.map((user: CustomerRepresentativeInfo) => ({
       id: user.id,
       companyname: user.companyname,
       lastUpdater: user.username,
@@ -205,17 +212,20 @@ const filterTableData = async (
     return;
   }
   try {
+    // /CustomerRepresentative?pageSize=${数据大小}&page=${页数}
     const response = await apiClient.post<apiResponseCustomerRepresentative>(
-      "/CustomerRepresentative/filter",
+      `/CustomerRepresentative/filter?pageSize=${pageSize.value}&page=${currentPage.value}`,
       {
         searchValue,
         searchType,
       }
     );
-    const responseData: any = response!.data;
-    filterData.value = responseData.map((user: CustomerRepresentative) => ({
+    const responseData: any = response!.data.data;
+    // 页面数据长度
+    pageTotal.value = response!.data.total;
+    filterData.value = responseData.map((user: CustomerRepresentativeInfo) => ({
       id: user.id,
-      companyName: user.companyname,
+      companyname: user.companyname,
       lastUpdater: user.username,
       companyAddress: user.address,
       createTime: user.created_at,
@@ -225,32 +235,44 @@ const filterTableData = async (
       address: user.address,
       classification: user.groups.map((group) => group.group_name),
     }));
+    console.log(filterData);
   } catch (error) {
     console.log(error);
   }
 };
+
+const refreshDat = () => {};
 // 创建一个防抖函数 在输入框输入最后一个字 500毫秒之后执行 filterTableData函数
 const debouncedFunc = debounce(filterTableData, 500);
-//分页框
-const currentPage = ref<number>(1);
-// 显示表单数据行数
-const pageSize = ref<number>(10);
 // 定义当分页大小变化时
 const handleSizeChange = (val: number) => {
-  console.log(`每页${val}条数据`);
+  // 如果有查询数据则修改查询数据的表单数据
+  if (searchValue.value == "") {
+    getTableData();
+  } else {
+    filterTableData(searchValue.value, searchOptionChoosed.value);
+  }
 };
+
 // 定义当页码变化时
 const handleCurrentChange = (val: number) => {
-  console.log(`当前在第${val}页`);
+  // 如果有查询数据则修改查询数据的表单数据
+  if (searchValue.value == "") {
+    getTableData();
+  } else {
+    filterTableData(searchValue.value, searchOptionChoosed.value);
+  }
 };
+
 const btnLoading = ref<boolean>(false);
 // 定义编辑操作
-const handleEdit = async (index: number, row: CustomerRepresentative) => {
+const handleEdit = async (index: number, row: CustomerRepresentativeInfo) => {
   btnLoading.value = true;
   const userRes = await apiClient.get<apiResponseUser & apiResponseData>(
     "/CustomerRepresentative/" + row.id
   );
   const customerInfo = userRes!.data;
+  console.log(customerInfo);
   // 获取组信息
   const groupRes = await apiClient.get<apiResponseData>(
     "/group/name/" + row.id
@@ -271,28 +293,72 @@ const handleEdit = async (index: number, row: CustomerRepresentative) => {
 };
 
 // 定义删除操作
-const handleDelete = async (index: number, row: CustomerRepresentative) => {
-  // 传入需要删除的数据
-  dialogInfo.value.data.companyname = row.companyname;
-  delUrl.value = "CustomerRepresentative";
-  dialogInfo.value.id = row.id;
-  confirmDelete.value = true;
+const handleDelete = async (index: number, row: CustomerRepresentativeInfo) => {
+  ElMessageBox.confirm(
+    `是否要删除第${index + 1}行、公司名为${row.companyname}?`,
+    "是否确认删除",
+    {
+      distinguishCancelAndClose: true,
+      confirmButtonText: "删除",
+      cancelButtonText: "取消",
+      type: "warning",
+    }
+  )
+    .then(async () => {
+      const response = await apiClient.delete<apiResponseData>(
+        "/CustomerRepresentative/" + row.id
+      );
+      // 在页面上展示删除成功的提示消息
+      ElMessage({
+        type: "success",
+        message: `删除了公司：${dialogInfo.value.data.companyname}`,
+      });
+      if (searchValue.value == "") {
+        getTableData();
+      } else {
+        filterTableData(searchValue.value, searchOptionChoosed.value);
+      }
+    })
+    .catch((action: Action) => {
+      console.log(action);
+      ElMessage({
+        type: "info",
+        message: action === "cancel" ? "取消删除" : "删除失败",
+      });
+    });
 };
 
 const multipleTableRef = ref<InstanceType<typeof ElTable>>();
 const multipleSelection = ref<any>([]);
 // 点击了批量删除
 const deleteMany = async () => {
-  const ids = multipleSelection.value.map((val: any) => val.id);
-  const promises = ids.map((id: any) =>
-    apiClient.delete<apiResponseUser & apiResponseData>(
-      "/CustomerRepresentative/" + id
-    )
-  );
-  const res = apiClient.all(promises);
-  console.log(res);
-  getTableData();
+  ElMessageBox.confirm(`是否要执行批量删除的操作`, "是否确认删除", {
+    distinguishCancelAndClose: true,
+    confirmButtonText: "删除",
+    cancelButtonText: "取消",
+    type: "warning",
+  })
+    .then(async () => {
+      const ids = multipleSelection.value.map((val: any) => val.id);
+      const promises = ids.map((id: any) =>
+        apiClient.delete<apiResponseUser & apiResponseData>(
+          "/CustomerRepresentative/" + id
+        )
+      );
+      if (searchValue.value == "") {
+        getTableData();
+      } else {
+        filterTableData(searchValue.value, searchOptionChoosed.value);
+      }
+    })
+    .catch((action: Action) => {
+      ElMessage({
+        type: "info",
+        message: action === "cancel" ? "取消删除" : "删除失败",
+      });
+    });
 };
+
 // 获取已点击的按钮数据
 const handleSelectionChange = (val: any) => {
   multipleSelection.value = val;
@@ -310,8 +376,8 @@ onMounted(() => {
   getTableData();
 });
 
+// 使用 watchEffect 监听搜索字段（关键字和搜索类型）变化并过滤表格数据
 watchEffect(() => {
-  // 使用 watchEffect 监听搜索字段（关键字和搜索类型）变化并过滤表格数据
   debouncedFunc(searchValue.value, searchOptionChoosed.value);
 });
 </script>
@@ -323,6 +389,7 @@ $views-li: 40px;
   column-gap: 10px;
   height: calc(100% - 120px);
   margin: 10px;
+
   .OrderList {
     width: 100%;
     height: 100%;
@@ -348,9 +415,11 @@ $views-li: 40px;
         margin-right: 10px;
       }
     }
+
     .table {
       height: 600px;
     }
+
     .demo-pagination-block {
       display: flex;
       align-items: center;
